Move and rename GraphQL types
